refactor(event): Make the sender just take `Meta` instead
This let's us avoid a round trip to a cloud event for the k8s types
