+++ pkg/event/event_test.go
@@ -12,101 +12,6 @@ import (
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
 )
 
-func TestSource(t *testing.T) {
-	testCases := map[string]struct {
-		namespace string
-		kind      string
-		name      string
-		expected  string
-	}{
-		"standard case": {
-			namespace: "test-namespace",
-			kind:      "Promotion",
-			name:      "test-promotion",
-			expected:  "test-namespace/Promotion/test-promotion",
-		},
-		// Currently don't need any others but leaving as a test table just in case
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			result := Source(tc.namespace, tc.kind, tc.name)
-			require.Equal(t, tc.expected, result)
-		})
-	}
-}
-
-func TestToCloudEvent(t *testing.T) {
-	// NOTE(thomastaylor312): This is mostly meant as a smoke test to make sure something isn't
-	// seriously broken with the event conversion logic. If we have any other issues, we can test
-	// every event
-	testCases := map[string]struct {
-		eventMeta Meta
-		// Because this is generic, we need to provide an empty type to deserialize back into
-		eventParsed    Meta
-		expectedType   string
-		expectedSource string
-	}{
-		"promotion event": {
-			eventMeta: &PromotionCreated{
-				Common: Common{
-					Project: "test-project",
-					Actor:   ptr.To("test-actor"),
-				},
-				Promotion: Promotion{
-					Name:       "test-promotion",
-					StageName:  "test-stage",
-					CreateTime: time.Date(2024, 1, 1, 12, 0, 0, 0, time.UTC),
-					Freight: &Freight{
-						Name: "test-freight",
-					},
-				},
-			},
-			eventParsed:    &PromotionCreated{},
-			expectedType:   EventTypePrefix + string(kargoapi.EventTypePromotionCreated),
-			expectedSource: "test-project/Promotion/test-promotion",
-		},
-		"freight event": {
-			eventMeta: &FreightApproved{
-				Common: Common{
-					Project: "test-project",
-					Actor:   ptr.To("test-actor"),
-				},
-				Freight: Freight{
-					Name:       "test-freight",
-					CreateTime: time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC),
-				},
-			},
-			eventParsed:    &FreightApproved{},
-			expectedType:   EventTypePrefix + string(kargoapi.EventTypeFreightApproved),
-			expectedSource: "test-project/Freight/test-freight",
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			cloudEvent := ToCloudEvent(tc.eventMeta)
-
-			// Verify event type
-			require.Equal(t, tc.expectedType, cloudEvent.Type())
-
-			// Verify source
-			require.Equal(t, tc.expectedSource, cloudEvent.Source())
-
-			// Verify ID is set (should be a UUID)
-			require.NotEmpty(t, cloudEvent.ID())
-
-			// Verify time is set
-			require.False(t, cloudEvent.Time().IsZero())
-
-			// Verify data is set
-			err := cloudEvent.DataAs(&tc.eventParsed)
-			require.NoError(t, err)
-			require.Equal(t, tc.eventMeta, tc.eventParsed, "Should get the same data after round trip")
-		})
-	}
-}
-
 func TestCommon_MarshalAnnotationsTo(t *testing.T) {
 	testCases := map[string]struct {
 		common   Common
