+++ pkg/event/kubernetes/sender.go
@@ -4,9 +4,10 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"strings"
+	"reflect"
+	"slices"
 
-	cloudevent "github.com/cloudevents/sdk-go/v2"
+	"github.com/fatih/structtag"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/tools/record"
 
@@ -26,12 +27,7 @@ func NewEventSender(recorder record.EventRecorder) *EventSender {
 }
 
 // Send sends the event to the Kubernetes EventRecorder
-func (s *EventSender) Send(_ context.Context, evt cloudevent.Event) error {
-	eventType := evt.Type()
-	if !strings.HasPrefix(eventType, event.EventTypePrefix) {
-		return fmt.Errorf("event type %s does not match expected prefix %s", eventType, event.EventTypePrefix)
-	}
-	eventType = strings.TrimPrefix(eventType, event.EventTypePrefix)
+func (s *EventSender) Send(_ context.Context, evt event.Meta) error {
 	// First up, extract event data. If it is our known event type, we can deserialize, if it
 	// doesn't match, we do a generic conversion to annotations
 	var annotations map[string]string
@@ -40,36 +36,16 @@ func (s *EventSender) Send(_ context.Context, evt cloudevent.Event) error {
 	// NOTE(thomastaylor312): My kingdom for an enum here so we would know if we exhausted all
 	// cases. Whenever we add a new type, it should be added here or it will fall into default
 	// handling
-	switch kargoapi.EventType(eventType) {
-	case kargoapi.EventTypePromotionCreated:
-		annotations, message, err = toAnnotationsAndMessage[*event.PromotionCreated](&evt)
-	case kargoapi.EventTypePromotionSucceeded:
-		annotations, message, err = toAnnotationsAndMessage[*event.PromotionSucceeded](&evt)
-	case kargoapi.EventTypePromotionFailed:
-		annotations, message, err = toAnnotationsAndMessage[*event.PromotionFailed](&evt)
-	case kargoapi.EventTypePromotionErrored:
-		annotations, message, err = toAnnotationsAndMessage[*event.PromotionErrored](&evt)
-	case kargoapi.EventTypePromotionAborted:
-		annotations, message, err = toAnnotationsAndMessage[*event.PromotionAborted](&evt)
-	case kargoapi.EventTypeFreightApproved:
-		annotations, message, err = toAnnotationsAndMessage[*event.FreightApproved](&evt)
-	case kargoapi.EventTypeFreightVerificationSucceeded:
-		annotations, message, err = toAnnotationsAndMessage[*event.FreightVerificationSucceeded](&evt)
-	case kargoapi.EventTypeFreightVerificationFailed:
-		annotations, message, err = toAnnotationsAndMessage[*event.FreightVerificationFailed](&evt)
-	case kargoapi.EventTypeFreightVerificationErrored:
-		annotations, message, err = toAnnotationsAndMessage[*event.FreightVerificationErrored](&evt)
-	case kargoapi.EventTypeFreightVerificationAborted:
-		annotations, message, err = toAnnotationsAndMessage[*event.FreightVerificationAborted](&evt)
-	case kargoapi.EventTypeFreightVerificationInconclusive:
-		annotations, message, err = toAnnotationsAndMessage[*event.FreightVerificationInconclusive](&evt)
-	case kargoapi.EventTypeFreightVerificationUnknown:
-		annotations, message, err = toAnnotationsAndMessage[*event.FreightVerificationUnknown](&evt)
-	default:
-		if evt.DataContentType() != cloudevent.ApplicationJSON {
-			return fmt.Errorf("unsupported content type %s for event type %s", evt.DataContentType(), eventType)
+	if slices.Contains(event.KnownEventTypes, evt.Type()) {
+		// Type assert that the event is the type we expect for Kubernetes. If it is a known event
+		// type, it should implement this interface
+		typedEvt, ok := evt.(unionedInterface)
+		if !ok {
+			return fmt.Errorf("failed to assert event type, this is programmer error: %T", evt)
 		}
-		annotations, err = convertToAnnotations(evt.Data())
+		annotations, message = toAnnotationsAndMessage(typedEvt)
+	} else {
+		annotations, err = convertToAnnotations(evt)
 		if err != nil {
 			return fmt.Errorf("failed to convert event data to annotations: %w", err)
 		}
@@ -83,19 +59,13 @@ func (s *EventSender) Send(_ context.Context, evt cloudevent.Event) error {
 	if err != nil {
 		return fmt.Errorf("failed to extract event data: %w", err)
 	}
-
-	// Now that we have the data and the type, put together an object reference for the event
-	parts := strings.Split(evt.Source(), "/")
-	if len(parts) != 3 {
-		return fmt.Errorf("invalid event source: %s", evt.Source())
-	}
 	reference := corev1.ObjectReference{
-		Namespace: parts[0],
-		Kind:      parts[1],
-		Name:      parts[2],
+		Namespace: evt.GetProject(),
+		Kind:      evt.Kind(),
+		Name:      evt.GetName(),
 	}
 	// Use the recorder to send the event
-	s.recorder.AnnotatedEventf(&reference, annotations, corev1.EventTypeNormal, eventType, message)
+	s.recorder.AnnotatedEventf(&reference, annotations, corev1.EventTypeNormal, string(evt.Type()), message)
 
 	return nil
 }
@@ -103,35 +73,81 @@ func (s *EventSender) Send(_ context.Context, evt cloudevent.Event) error {
 type unionedInterface interface {
 	event.AnnotationMarshaler
 	event.Message
+	event.Meta
 }
 
-func toAnnotationsAndMessage[T unionedInterface](evt *cloudevent.Event) (map[string]string, string, error) {
-	var data T
-	if err := evt.DataAs(&data); err != nil {
-		return nil, "", fmt.Errorf("failed to unmarshal event data: %w", err)
-	}
-	message := data.GetMessage()
-	annotations := data.MarshalAnnotations()
+func toAnnotationsAndMessage(evt unionedInterface) (map[string]string, string) {
+	message := evt.GetMessage()
+	annotations := evt.MarshalAnnotations()
 
-	return annotations, message, nil
+	return annotations, message
 }
 
-// convertToAnnotations takes raw JSON data and decodes it to a `map[string]interface{}`. It then
-// converts raw types to string and JSON encodes anything else that isn't a raw type (e.g. maps,
-// slices, structs). It will take each data key and prefix it with
-// `kargoapi.AnnotationEventKeyPrefix` and then return the map.
-func convertToAnnotations(data []byte) (map[string]string, error) {
-	var rawData map[string]any
-	if err := json.Unmarshal(data, &rawData); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal event data: %w", err)
-	}
+// convertToAnnotations uses reflection to iterate over the fields of the given struct and converts
+// them to annotations. It skips any fields that have nil values. It will take each field name and
+// prefix it with `kargoapi.AnnotationEventKeyPrefix` and then return the map.
+func convertToAnnotations(data event.Meta) (map[string]string, error) {
 
+	// NOTE(thomastaylor312): I tried to do this
 	annotations := make(map[string]string)
-	for key, value := range rawData {
-		annotationKey := kargoapi.AnnotationKeyEventPrefix + key
-		annotationValue, err := convertValueToString(value)
+
+	rv := reflect.ValueOf(data)
+	rt := reflect.TypeOf(data)
+
+	// Handle pointer to struct
+	if rv.Kind() == reflect.Pointer {
+		if rv.IsNil() {
+			return annotations, nil
+		}
+		rv = rv.Elem()
+		rt = rt.Elem()
+	}
+
+	// Ensure we're working with a struct
+	if rv.Kind() != reflect.Struct {
+		return nil, fmt.Errorf("expected struct, got %s", rv.Kind())
+	}
+
+	// Iterate over struct fields
+	for i := 0; i < rv.NumField(); i++ {
+		field := rv.Field(i)
+		fieldType := rt.Field(i)
+
+		// Skip unexported fields
+		if !fieldType.IsExported() {
+			continue
+		}
+
+		// Shouldn't happen, but for safety sake we check
+		if !field.IsValid() {
+			continue
+		}
+
+		// Skip nil pointer fields
+		if field.Kind() == reflect.Pointer && field.IsNil() {
+			continue
+		}
+
+		// Skip nil interface fields
+		if field.Kind() == reflect.Interface && field.IsNil() {
+			continue
+		}
+
+		val := field.Interface()
+
+		// Unalias any string types for this next step so they are processed properly
+		if field.Kind() == reflect.String {
+			val = field.String()
+		}
+
+		// Get the field name from the JSON tag or the struct field name
+		fieldName := getFieldName(fieldType)
+
+		// Convert field value to string
+		annotationKey := kargoapi.AnnotationKeyEventPrefix + fieldName
+		annotationValue, err := convertValueToString(val)
 		if err != nil {
-			return nil, fmt.Errorf("failed to convert value for key %s: %w", key, err)
+			return nil, fmt.Errorf("failed to convert value for field %s: %w", fieldName, err)
 		}
 		annotations[annotationKey] = annotationValue
 	}
@@ -145,6 +161,12 @@ func convertValueToString(value any) (string, error) {
 	switch v := value.(type) {
 	case string:
 		return v, nil
+	case *string:
+		if v != nil {
+			return *v, nil
+		}
+		return "", nil
+	// Special cases: the Event
 	default:
 		// For complex types (maps, slices, structs), JSON encode them
 		jsonBytes, err := json.Marshal(v)
@@ -153,4 +175,17 @@ func convertValueToString(value any) (string, error) {
 		}
 		return string(jsonBytes), nil
 	}
+}
+
+// Get field name from the JSON tag or the struct field name
+func getFieldName(field reflect.StructField) string {
+	tags, err := structtag.Parse(string(field.Tag))
+	if err != nil {
+		return field.Name
+	}
+	jsonTag, err := tags.Get("json")
+	if err != nil {
+		return field.Name
+	}
+	return jsonTag.Name
 }
